+++ internal/core/service/service.go
@@ -28,7 +28,7 @@ func NewDefaultRestaurantService(restaurantRepository port.RestaurantRepository,
 	return &DefaultRestaurantService{restaurantRepository: restaurantRepository, domainEventPublisher: domainEventPublisher, trManager: trManager}
 }
 
-func (rs *DefaultRestaurantService) FindAll(ctx context.Context, offset uint32, limit uint8) ([]*domain.Restaurant, error) {
+func (rs *DefaultRestaurantService) FindAll(ctx context.Context, offset int, limit int) ([]*domain.Restaurant, error) {
 	restaurants, err := rs.restaurantRepository.FindAll(ctx, offset, limit)
 	if err != nil {
 		log.Error().Msg("an error occurred while fetching all the restaurant: " + err.Error())
